Job and Step definitions are no longer in massive source file
